agent/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
httpx==0.25.0
pydantic==2.4.2
PyYAML==6.0.1
python-multipart==0.0.6
jinja2==3.1.2
